+++ src/arche/tools/schema.py
@@ -1,7 +1,4 @@
 from collections import defaultdict
-from dataclasses import dataclass
-import json
-import pprint
 import random
 from typing import Any, Deque, Dict, List, Optional
 
@@ -16,18 +13,7 @@ import pandas as pd
 from tqdm import tqdm_notebook
 
 
-@dataclass
-class BasicSchema:
-    d: Schema
-
-    def json(self):
-        print(json.dumps(self.d, indent=4))
-
-    def __repr__(self):
-        return pprint.pformat(self.d)
-
-
-def basic_json_schema(data_source: str, items_numbers: List[int] = None) -> BasicSchema:
+def basic_json_schema(data_source: str, items_numbers: List[int] = None) -> Schema:
     """Print a json schema based on the provided job_key and item numbers
 
     Args:
@@ -35,7 +21,7 @@ def basic_json_schema(data_source: str, items_numbers: List[int] = None) -> Basi
         items_numbers: array of item numbers to create schema from
     """
     schema = create_json_schema(data_source, items_numbers)
-    return BasicSchema(schema)
+    return Schema(schema)
 
 
 def create_json_schema(